Add files via upload
This version (v4) resolves issues with the for loop and if statements that iteratively open, read, and write to the data frame. Key error was bad syntax - did not intent lines of code after a for statement which made the program skip certain files and result in a conflict with the index.